+++ src/GRA.Domain.Repository/IPageRepository.cs
@@ -6,9 +6,10 @@ namespace GRA.Domain.Repository
 {
     public interface IPageRepository : IRepository<Page>
     {
-        Task<Page> GetByStubAsync(int siteId, string pageStub);
+        Task<Page> GetByStubAsync(int siteId, string pageStub, bool exlcudeDashboardPage);
         Task<IEnumerable<Page>> PageAllAsync(int siteId, int skip, int take);
         Task<int> GetCountAsync(int siteId);
         Task<IEnumerable<Page>> GetFooterPagesAsync(int siteId);
+        Task<Page> GetDashboardPageAsync(int siteId);
     }
 }